Fix inconsistent filename and log exception
